backend/requirements-prod.txt
# FastAPI Enterprise Boilerplate - Production Dependencies
# Python 3.13+
# Last Updated: February 2026
# NOTE: Dev/test dependencies live in requirements.txt only.

# ============================================
# Core Framework
# ============================================
fastapi>=0.128.8,<1.0.0
uvicorn[standard]>=0.40.0
pydantic>=2.12.5,<3.0.0
pydantic-settings>=2.12.0

# ============================================
# Database
# ============================================
sqlalchemy[asyncio]>=2.0.46,<3.0.0
asyncpg>=0.31.0
alembic>=1.18.4
greenlet>=3.3.1

# ============================================
# Authentication & Security
# ============================================
PyJWT>=2.11.0
cryptography>=44.0.0
bcrypt>=4.2.1
python-multipart>=0.0.22

# ============================================
# Redis & Caching
# ============================================
redis>=5.2.1

# ============================================
# Observability
# ============================================
opentelemetry-api>=1.39.1
opentelemetry-sdk>=1.39.1
opentelemetry-instrumentation-fastapi>=0.60b1
opentelemetry-instrumentation-sqlalchemy>=0.60b1
opentelemetry-instrumentation-redis>=0.60b1
opentelemetry-instrumentation-httpx>=0.60b1
opentelemetry-exporter-otlp>=1.39.1
prometheus-client>=0.24.1

# ============================================
# Background Tasks
# ============================================
arq>=0.27.0

# ============================================
# CLI
# ============================================
typer[all]>=0.23.0
rich>=13.9.4

# ============================================
# Email
# ============================================
fastapi-mail>=1.6.1
jinja2>=3.1.6

# ============================================
# MFA / Two-Factor Authentication
# ============================================
pyotp>=2.9.0
qrcode[pil]>=8.2

# ============================================
# i18n / Internationalization
# ============================================
babel>=2.18.0

# ============================================
# Utilities
# ============================================
httpx>=0.28.1
python-dotenv>=1.2.1
email-validator>=2.3.0
aiofiles>=25.1.0

# ============================================
# Data Exchange - PDF/Excel
# ============================================
openpyxl>=3.1.5
weasyprint>=68.1
